Recon_tool/redrecon/requirements.txt
requests>=2.31.0
dnspython>=2.4.0
urllib3>=2.0.0
